feat: refactor models
